requirements.txt
python-decouple==3.3
pyModbusTCP==0.1.8

